+++ compiler/chem_clang.h
@@ -8,9 +8,9 @@ namespace clang {
 
 struct ErrorMsg {
     const char *filename_ptr; // can be null
-    size_t filename_len;
+    unsigned long filename_len;
     const char *msg_ptr;
-    size_t msg_len;
+    unsigned long msg_len;
     const char *source; // valid until the ASTUnit is freed. can be null
     unsigned line; // 0 based
     unsigned column; // 0 based
@@ -18,4 +18,4 @@ struct ErrorMsg {
 };
 
 clang::ASTUnit *ClangLoadFromCommandLine(const char **args_begin, const char **args_end,
-                                         struct ErrorMsg **errors_ptr, size_t *errors_len, const char *resources_path);
+                                         struct ErrorMsg **errors_ptr, unsigned long *errors_len, const char *resources_path);
